Days 10&11: fix tests; improve performance

+++ days/include/days/day10.hpp
@@ -2,4 +2,4 @@
 
 #include "days/day.hpp"
 
-using Day10 = Day<10, std::array<i8, 240>, i32, std::string_view>;
+using Day10 = Day<10, std::array<i32, 240>, i32, std::string_view>;
